pyproject.toml
[project]
name = "langtarder-agents"
version = "0.1.0"
description = "AI-powered cryptocurrency trading system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core trading
    "ccxt>=4.5.28",
    
    # AI/LLM
    "langchain>=0.3.0",
    "langchain-anthropic>=1.3.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.1.0",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langsmith>=0.5.0",
    
    # Data processing
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    
    # Database
    "psycopg2-binary>=2.9.11",
    "sqlmodel>=0.0.27",
    
    # API (FastAPI)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.6.0",
    
    # Utilities
    "jsonschema>=4.25.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
]

